requirements.txt
absl-py==2.4.0
array_record==0.8.3
astunparse==1.6.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.2
etils==1.13.0
flatbuffers==25.12.19
fsspec==2026.1.0
gast==0.7.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
immutabledict==4.2.2
importlib_resources==6.5.2
keras==3.13.2
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.1
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
promise==2.3
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.0
pygame==2.6.1
Pygments==2.19.2
requests==2.32.5
rich==14.3.1
scipy==1.17.0
setuptools==80.10.2
simple-parsing==0.1.8
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.3
tensorflow_cpu==2.20.0
termcolor==3.3.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.0.1
zipp==3.23.0
